move data table class to leaderboard components
