+++ lib/pages/leaderboard_page/components/leaderboard_table.dart
@@ -3,7 +3,7 @@ import 'package:dalal_street_client/constants/leaderboard_type.dart';
 import 'package:dalal_street_client/theme/colors.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:dalal_street_client/utils/paginated_data_table_2.dart';
+import 'package:dalal_street_client/pages/leaderboard_page/components/paginated_data_table_2.dart';
 import '../leaderboard_page.dart';
 
 class LeaderboardTable extends StatefulWidget {
